Fixed 263000: [Upload] Upload widget reports wrong size for getUploadTotal() https://bugs.eclipse.org/bugs/show_bug.cgi?id=263000
Reworked the transfer of progress data. These are generated on the server-side (provided by FileUpload lib) and were transfered to the Browser and back to the server (LCA) before. This has changed, so that the LCA reads the progress data from a server side state tracker, now.
Fixed another issue where a second readAndDispatch-Loop was spawned and never terminated.